requirements.txt
flask
psycopg2-binary
python-decouple
python-dotenv